requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
channels>=4.0,<5.0
channels-redis>=4.1,<5.0
daphne>=4.0,<5.0
python-dotenv>=1.0,<2.0
psycopg2-binary>=2.9,<3.0
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
pytest-asyncio>=0.21,<1.0
asgiref~=3.11.1